1.51 TZ11 - add FilmDbStorage - add FilmGenre class - add FilmGenresDb

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/GenreController.java
@@ -28,7 +28,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}")
-    public Genre findGenre(@PathVariable("id") Long id) {
+    public Genre findGenre(@PathVariable("id") Integer id) {
         log.info("get genre with id= " + id);
         return genreService.findGenre(id);
     }
